+++ app/src/main/java/com/example/bayar/cityweather/model/Wind.java
@@ -11,20 +11,20 @@ public class Wind {
 
     @SerializedName("speed")
     @Expose
-    Long speed;
+    Double speed;
 
     public Wind() {
     }
 
-    public Wind(Long speed) {
+    public Wind(Double speed) {
         this.speed = speed;
     }
 
-    public Long getSpeed() {
+    public Double getSpeed() {
         return speed;
     }
 
-    public void setSpeed(Long speed) {
+    public void setSpeed(Double speed) {
         this.speed = speed;
     }
 }
